Set git pull behavior to normal default (i.e., merge). See msg.
On recent versions of `git`, unless this is set, it causes a
warning. I don't want to use any of the newer options, so I am setting
it to the old default (i.e., try to fast forward on pull, if that
fails, then try to merge); using `rebase` is quite dangerous so I
definitely don't want that as my global default. Using ff only as
default is safer than the merge default, but I am already used to
cancelling merges when I want, and otherwise just reusing the merge
that happens, so this default makes more sense.